Correção do tipo das colunas preço e valor para decimal

+++ source/laravel/database/migrations/2023_12_22_190127_create_cars_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->unsignedBigInteger('color_id');
             $table->integer('doors');
             $table->integer('year');
-            $table->float('km');
-            $table->float('price');
+            $table->decimal('km', 10, 2);
+            $table->decimal('price', 10, 2);
             $table->timestamps();
         });
     }
